Create base entity for SimpliSafe
