+++ homeassistant/components/simplisafe/__init__.py
@@ -22,7 +22,11 @@ from homeassistant.helpers import (
     config_validation as cv,
     device_registry as dr,
 )
-from homeassistant.helpers.dispatcher import async_dispatcher_send
+from homeassistant.helpers.dispatcher import (
+    async_dispatcher_connect,
+    async_dispatcher_send,
+)
+from homeassistant.helpers.entity import Entity
 from homeassistant.helpers.event import async_track_time_interval
 from homeassistant.helpers.service import verify_domain_control
 
@@ -240,3 +244,58 @@ class SimpliSafe:
         tasks = [self._update_system(system) for system in self.systems.values()]
 
         await asyncio.gather(*tasks)
+
+
+class SimpliSafeEntity(Entity):
+    """Define a base SimpliSafe entity."""
+
+    def __init__(self, system):
+        """Initialize."""
+        self._system = system
+        self._async_unsub_dispatcher_connect = None
+        self._attrs = {ATTR_SYSTEM_ID: system.system_id}
+
+    @property
+    def device_info(self):
+        """Return device registry information for this entity."""
+        return {
+            "identifiers": {(DOMAIN, self._system.system_id)},
+            "manufacturer": "SimpliSafe",
+            "model": self._system.version,
+            # The name should become more dynamic once we deduce a way to
+            # get various other sensors from SimpliSafe in a reliable manner:
+            "name": "Keypad",
+            "via_device": (DOMAIN, self._system.serial),
+        }
+
+    @property
+    def device_state_attributes(self):
+        """Return the state attributes."""
+        return self._attrs
+
+    @property
+    def name(self):
+        """Return the name of the entity."""
+        return self._system.address
+
+    @property
+    def unique_id(self):
+        """Return the unique ID of the entity."""
+        return self._system.system_id
+
+    async def async_added_to_hass(self):
+        """Register callbacks."""
+
+        @callback
+        def update():
+            """Update the state."""
+            self.async_schedule_update_ha_state(True)
+
+        self._async_unsub_dispatcher_connect = async_dispatcher_connect(
+            self.hass, TOPIC_UPDATE, update
+        )
+
+    async def async_will_remove_from_hass(self) -> None:
+        """Disconnect dispatcher listener when removed."""
+        if self._async_unsub_dispatcher_connect:
+            self._async_unsub_dispatcher_connect()
